requirements.txt
Pygments==1.6
argparse==1.2.1
-e git+git@github.com:koddsson/coredata-python-client.git@f5816f9d8d63dfbf054922550fa691f876f8cc91#egg=coredata_python_client-master
docopt==0.6.1
enum34==1.0
fancycompleter==0.4
fusepy==2.0.2
pdbpp==0.7.2
pyrepl==0.8.4
requests==2.3.0
wmctrl==0.1
wsgiref==0.1.2
